Move execute commands into clowder pool to allow killing subprocesses
