+++ clowder/utility/git_utilities.py
@@ -6,8 +6,8 @@ import subprocess
 import sys
 from git import Repo, GitError
 from termcolor import colored, cprint
+from clowder.utility.clowder_pool import ClowderPool
 from clowder.utility.clowder_utilities import (
-    execute_command,
     existing_git_repository,
     is_offline,
     ref_type,
@@ -132,7 +132,7 @@ class Git(object):
                 message = colored(' - Failed to fetch from ', 'red')
                 error = message + remote_output + ' ' + ref_output
                 command = ['git', 'fetch', remote, truncate_ref(ref), '--depth', str(depth), '--prune', '--tags']
-        return_code = execute_command(command, self.repo_path)
+        return_code = ClowderPool.execute_command(command, self.repo_path)
         if return_code != 0:
             print(error)
             if remove_dir:
@@ -270,7 +270,7 @@ class Git(object):
             command = ['git', 'branch', '-r']
         else:
             return
-        return_code = execute_command(command, self.repo_path)
+        return_code = ClowderPool.execute_command(command, self.repo_path)
         if return_code != 0:
             cprint(' - Failed to print branches', 'red')
             print_command_failed_error(command)
@@ -411,7 +411,7 @@ class Git(object):
             self._pull(upstream_remote, truncate_ref(ref))
         print(' - Push to ' + fork_remote_output + ' ' + branch_output)
         command = ['git', 'push', fork_remote, truncate_ref(ref)]
-        return_code = execute_command(command, self.repo_path)
+        return_code = ClowderPool.execute_command(command, self.repo_path)
         if return_code != 0:
             message = colored(' - Failed to push to ', 'red')
             print(message + fork_remote_output + ' ' + branch_output)
@@ -820,7 +820,7 @@ class Git(object):
         remote_output = format_remote_string(remote)
         print(' - Pull from ' + remote_output + ' ' + branch_output)
         command = ['git', 'pull', remote, branch]
-        return_code = execute_command(command, self.repo_path)
+        return_code = ClowderPool.execute_command(command, self.repo_path)
         if return_code != 0:
             message = colored(' - Failed to pull from ', 'red')
             print(message + remote_output + ' ' + branch_output)
@@ -836,7 +836,7 @@ class Git(object):
         remote_output = format_remote_string(remote)
         print(' - Rebase onto ' + remote_output + ' ' + branch_output)
         command = ['git', 'pull', '--rebase', remote, branch]
-        return_code = execute_command(command, self.repo_path)
+        return_code = ClowderPool.execute_command(command, self.repo_path)
         if return_code != 0:
             message = colored(' - Failed to rebase onto ', 'red')
             print(message + remote_output + ' ' + branch_output)
